FIX yang bener websitenya

+++ PA1/process_ongkir.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 session_start();
 $ongkos = isset($_POST['ongkir']) ? $_POST['ongkir'] : 0;
 $ongkos = str_replace(',', '', $ongkos); // Menghapus koma
@@ -20,26 +20,32 @@ $fileError = $file['error'];
 if ($fileError === UPLOAD_ERR_OK) {
     // Mendapatkan ekstensi file
     $fileExt = pathinfo($fileName, PATHINFO_EXTENSION);
-    // Menghasilkan nama unik untuk file
-    $newFileName = uniqid() . '.' . $fileExt;
-    // Lokasi penyimpanan file
-    $uploadPath = 'lampiran_keranjang/' . $newFileName;
+    $allowedExtensions = array('jpeg', 'jpg', 'png');
 
-    // Pindahkan file ke lokasi penyimpanan
-    if (move_uploaded_file($fileTmp, $uploadPath)) {
-        $hapusChartData = "UPDATE chart_data SET status = 'include ongkir' WHERE id_chartdata IN (SELECT id_chartdata FROM chart WHERE id_user = '$id_user')";
-        $hapusChart = "DELETE FROM chart WHERE id_user = '$id_user'";
-        $query_sql = "INSERT INTO checkout_produk (id_checkout,total_pembayaran,tanggal_pemesanan, status_bayar, bukti_bayar, akun_id,lampiran_keranjang) VALUES ('','$hasil','$tanggal','Bayar Sekarang','', '$id_user','$newFileName')";
+    if (in_array(strtolower($fileExt), $allowedExtensions)) {
+        // Menghasilkan nama unik untuk file
+        $newFileName = uniqid() . '.' . $fileExt;
+        // Lokasi penyimpanan file
+        $uploadPath = 'lampiran_keranjang/' . $newFileName;
 
-        if (mysqli_query($conn, $hapusChartData) && mysqli_query($conn, $hapusChart) && mysqli_query($conn, $query_sql)) {
-            echo "<script>alert('Checkout Produk Berhasil Silahkan Melakukan Pembayaran.'); window.location.href = 'profile_user.php';</script>";
+        // Pindahkan file ke lokasi penyimpanan
+        if (move_uploaded_file($fileTmp, $uploadPath)) {
+            $hapusChartData = "UPDATE chart_data SET status = 'include ongkir' WHERE id_chartdata IN (SELECT id_chartdata FROM chart WHERE id_user = '$id_user')";
+            $hapusChart = "DELETE FROM chart WHERE id_user = '$id_user'";
+            $query_sql = "INSERT INTO checkout_produk (id_checkout,total_pembayaran,tanggal_pemesanan, status_bayar, bukti_bayar, akun_id,lampiran_keranjang) VALUES ('','$hasil','$tanggal','Bayar Sekarang','', '$id_user','$newFileName')";
+
+            if (mysqli_query($conn, $hapusChartData) && mysqli_query($conn, $hapusChart) && mysqli_query($conn, $query_sql)) {
+                echo "<script>alert('Checkout Produk Berhasil Silahkan Melakukan Pembayaran.'); window.location.href = 'profile_user.php';</script>";
+            } else {
+                echo "<script>alert('Pembelian produk gagal: " . mysqli_error($conn) . "');</script>";
+            }
         } else {
-            echo "<script>alert('Pembelian produk gagal: " . mysqli_error($conn) . "');</script>";
+            echo "<script>alert('Gagal mengunggah gambar lampiran keranjang');</script>";
         }
     } else {
-        echo "<script>alert('Gagal mengunggah gambar lampiran keranjang');</script>";
+        echo "<script>alert('File yang diunggah harus dalam format JPEG, JPG, atau PNG.'); window.location.href = 'cekongkir.php';</script>";
     }
 } else {
-    echo "<script>alert('Terjadi kesalahan saat mengunggah gambar lampiran keranjang');</script>";
+    echo "<script>alert('Terjadi kesalahan saat mengunggah gambar lampiran keranjang'); window.location.href = 'cekongkir.php';</script>";
 }
 ?>
